# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me/addresses": {
      "get": {
        "operationId": "listMyAddresses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adresses sauvegardées",
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "createMyAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajoute une adresse au carnet",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/addresses/{id}": {
      "put": {
        "operationId": "updateMyAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remplace une adresse du carnet",
        "tags": [
          "account"
        ]
      },
      "delete": {
        "operationId": "deleteMyAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime une adresse du carnet",
        "tags": [
          "account"
        ]
      }
    },
...
    "/v1/cart": {
      "get": {
        "description": "Jeton invité inconnu ou expiré ⇒ panier vide (jamais d’erreur). Les écarts corrigés (produit dépublié, épuisé, prix changé) sont signalés une seule fois dans `changes`.",
        "operationId": "getCart",
        "parameters": [
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Jeton de panier invité (émis à la première addition ou via POST /v1/auth/guest-cart). Ignoré quand un Bearer est présent : le panier du compte prime.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          }
        },
        "summary": "Panier courant, revalidé côté serveur",
        "tags": [
          "cart"
        ]
      }
    },
    "/v1/cart/items": {
      "post": {
        "description": "Sans panier existant, un panier est créé ; pour un invité, `guestCartToken` est retourné UNE seule fois — à stocker et renvoyer dans X-Cart-Token.",
        "operationId": "addCartItem",
        "parameters": [
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Jeton de panier invité (émis à la première addition ou via POST /v1/auth/guest-cart). Ignoré quand un Bearer est présent : le panier du compte prime.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          }
        },
        "summary": "Ajoute une variante (les quantités s’additionnent)",
        "tags": [
          "cart"
        ]
      }
    },
    "/v1/cart/items/{variantId}": {
      "patch": {
        "operationId": "updateCartItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Jeton de panier invité (émis à la première addition ou via POST /v1/auth/guest-cart). Ignoré quand un Bearer est présent : le panier du compte prime.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          }
        },
        "summary": "Fixe la quantité d’une ligne",
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "operationId": "removeCartItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Jeton de panier invité (émis à la première addition ou via POST /v1/auth/guest-cart). Ignoré quand un Bearer est présent : le panier du compte prime.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          }
        },
        "summary": "Retire une ligne",
        "tags": [
          "cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "FFC API",
    "description": "API de la plateforme Filtration Montréal / Furnace Filters Canada",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "enum": [
              "CA",
              "US"
            ]
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isDefaultShipping": {
            "type": "boolean"
          },
          "isDefaultBilling": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "line1",
          "city",
          "province",
          "postalCode",
          "country",
          "isDefaultShipping",
          "isDefaultBilling"
        ]
      },
      "AddressListDto": {
        "type": "object",
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            }
          }
        },
        "required": [
          "addresses"
        ]
      },
      "SaveAddressDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          },
          "company": {
            "type": "string",
            "maxLength": 150
          },
          "line1": {
            "type": "string",
            "maxLength": 200
          },
          "line2": {
            "type": "string",
            "maxLength": 200
          },
          "city": {
            "type": "string",
            "maxLength": 120
          },
          "province": {
            "type": "string",
            "description": "Code de province (CA) ou d’état (US)",
            "example": "QC"
          },
          "postalCode": {
            "type": "string",
            "description": "Code postal (A1A 1A1) ou ZIP (12345[-6789])",
            "maxLength": 12
          },
          "country": {
            "type": "string",
            "enum": [
              "CA",
              "US"
            ]
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "label": {
            "type": "string",
            "maxLength": 60,
            "description": "Libellé libre (« Maison », « Chalet »…)"
          },
          "isDefaultShipping": {
            "type": "boolean"
          },
          "isDefaultBilling": {
            "type": "boolean"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "line1",
          "city",
          "province",
          "postalCode",
          "country"
        ]
      },
...
      "CartLineDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPriceCents": {
            "type": "number",
            "description": "Prix courant de la variante (cents) — recalculé à chaque lecture"
          },
          "currency": {
            "type": "string",
            "enum": [
              "CAD",
              "USD"
            ]
          },
          "lineSubtotalCents": {
            "type": "number",
            "description": "quantity × unitPriceCents"
          },
          "nameFr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "slugFr": {
            "type": "string",
            "nullable": true
          },
          "slugEn": {
            "type": "string",
            "nullable": true
          },
          "nominalLabel": {
            "type": "string",
            "nullable": true,
            "example": "16x25x1"
          },
          "packSize": {
            "type": "number"
          },
          "merv": {
            "type": "number",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "availableQuantity": {
            "type": "number",
            "description": "Stock vendable restant (en main − réservé)"
          }
        },
        "required": [
          "variantId",
          "productId",
          "sku",
          "quantity",
          "unitPriceCents",
          "currency",
          "lineSubtotalCents",
          "nameFr",
          "nameEn",
          "packSize",
          "availableQuantity"
        ]
      },
      "RemovedCartLineDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "nameFr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "UNAVAILABLE",
              "OUT_OF_STOCK"
            ],
            "description": "UNAVAILABLE : produit retiré du catalogue · OUT_OF_STOCK : épuisé"
          }
        },
        "required": [
          "sku",
          "nameFr",
          "nameEn",
          "reason"
        ]
      },
      "AdjustedCartLineDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "nameFr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "fromQuantity": {
            "type": "number",
            "description": "Quantité demandée avant ajustement"
          },
          "toQuantity": {
            "type": "number",
            "description": "Quantité retenue (stock restant)"
          }
        },
        "required": [
          "sku",
          "nameFr",
          "nameEn",
          "fromQuantity",
          "toQuantity"
        ]
      },
      "PriceChangedCartLineDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "nameFr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "fromCents": {
            "type": "number",
            "description": "Prix consigné à l’ajout (cents)"
          },
          "toCents": {
            "type": "number",
            "description": "Prix courant (cents)"
          }
        },
        "required": [
          "sku",
          "nameFr",
          "nameEn",
          "fromCents",
          "toCents"
        ]
      },
      "CartChangesDto": {
        "type": "object",
        "properties": {
          "removed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemovedCartLineDto"
            }
          },
          "adjusted": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdjustedCartLineDto"
            }
          },
          "priceChanged": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceChangedCartLineDto"
            }
          }
        },
        "required": [
          "removed",
          "adjusted",
          "priceChanged"
        ]
      },
      "CartDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Null tant qu’aucun panier n’existe pour ce client"
          },
          "guestCartToken": {
            "type": "string",
            "nullable": true,
            "description": "Jeton de panier invité — présent UNIQUEMENT à la création (première addition sans compte). À stocker côté client et à renvoyer dans X-Cart-Token."
          },
          "currency": {
            "type": "string",
            "enum": [
              "CAD",
              "USD"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLineDto"
            }
          },
          "subtotalCents": {
            "type": "number",
            "description": "Somme des lignes (cents) — les taxes se calculent au checkout"
          },
          "itemCount": {
            "type": "number",
            "description": "Nombre total d’unités"
          },
          "changes": {
            "$ref": "#/components/schemas/CartChangesDto"
          }
        },
        "required": [
          "currency",
          "items",
          "subtotalCents",
          "itemCount",
          "changes"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid",
            "description": "Variante de produit à ajouter"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 99,
            "example": 1
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 99
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}